main gained PR #76 (au-host capabilities) and PR #87 (vst2 param properties).
One conflict, plus one gap #87 makes closable.

## Conflict: tutti-au-host/src/parameters.rs

An add/add of two disjoint test sets into the same `mod tests` — this branch's
`Indexed` unit-decode tests, and #76's per-element addressing / clump / display
-curve tests. Resolved as a union, not a choice: both sides' final test ended
mid-assertion sharing an identical 7-line closing tail, so each side gets its
own copy of it rather than the two bodies being spliced together. Verified by
diffing the test-name sets that all 15 survive — 0 lost from either parent.

## Gap closed: VST2 opcode 56 now reaches the shared type

This branch's `parameter_list` documented `effGetParameterProperties` as "would
report a real range, and this is where it would land", because the vendored
crate surfaced no way to ask. #87 landed exactly that: `param_properties.rs`
with `Vst2Instance::parameter_properties(id) -> Option<ParameterProperties>`.

Left alone, the merge would compile clean while decoding the opcode and then
dropping it — every VST2 parameter still reported `Normalized`/`Unknown`. So
`parameter_list` now reads it per parameter.

The two designs compose because #87 reached the same conclusion independently:
its `integer_range: Option<IntegerRange>` is `Option` for the reason `ParamRange`
is a sum type — a bounds pair present only under `USES_INT_STEP` must not be
readable without handling its absence.

Three details the wiring has to get right:

- The gate is per-flag, not per-answer. `USES_INT_STEP` gates the range;
  `USES_FLOAT_STEP` gates a granularity carrying no bounds. A plugin declaring
  only the latter gets `Normalized` with no steps.
- `IntegerRange::step_count()` returns `None` for a declared-but-unwalkable
  range (non-positive step, inverted bounds). That maps to `ParamSteps::Unknown`,
  not `Continuous` — the plugin spoke, but not usably.
- `step_count()` counts steps *between* endpoints; `ParamSteps` counts positions.
  0..127 in steps of 1 is 127 steps and 128 positions.

Two tests added. The load-bearing half is negative: the probe leaves
`min_integer`/`max_integer` at 999/-999 on every parameter except the one that
declares a range, with only `USES_FLOAT_STEP` set — so a host that reads the
fields without checking the gate produces a `Plain` range of 999..-999. Verified
by mutation: dropping the wiring fails
`a_declared_integer_range_reaches_the_shared_parameter_info`.

## Verification

- Engine workspace `--all-targets --all-features` EXIT 0
- App workspace `-p dawai-wasm-plugin --all-targets` EXIT 0
- 97 vst2-host tests pass, 15 au-host parameter tests pass
- rustfmt clean; clippy clean outside vendored crates

Pre-existing and unchanged, verified on origin/main earlier: `tutti-midi-io`
(DiscoveryData fields) and `bevy-tutti` (UmpOutRes::new arity) fail lib tests.

+++ crates/plugin/tutti-plugin-server/src/loaders/au.rs
@@ -280,8 +280,11 @@ impl AuInstance {
 
             // AU exposes a single main bus per direction here.
             let loaded = LoadedPlugin {
-                inputs: single_bus(inner.num_inputs() as usize),
-                outputs: single_bus(inner.num_outputs() as usize),
+                // `num_inputs`/`num_outputs` are `u32` off the AU element
+                // count; `From<u32>` canonicalizes them, so the `as usize`
+                // hop is gone.
+                inputs: single_bus(inner.num_inputs()),
+                outputs: single_bus(inner.num_outputs()),
                 latency_samples: latency,
                 features,
             };
